API/requirements.txt
fastapi
uvicorn
mysql-connector-python
sqlalchemy
pydantic
bcrypt